Moved MockAerohiveFixture into source.
