+++ mock_aerohive/__init__.py
@@ -1,1 +1,2 @@
 from MockAerohive import MockAerohive
+from MockAerohiveFixture import MockAerohive as MockAerohiveFixture
