+++ test/integration/command_error_handling/test_unknown_keyword_hostname.py
@@ -1,6 +1,6 @@
 import pexpect
 
-from test.util.MockAerohiveFixture import MockAerohive
+import mock_aerohive.MockAerohiveFixture as MockAerohive
 
 def test_hostname_three_arguments(MockAerohive):
     username = "admin"
