+++ test/integration/start_stop/test_stop_kills_server.py
@@ -1,6 +1,6 @@
 import pexpect
 
-from test.util.MockAerohiveFixture import MockAerohive
+import mock_aerohive.MockAerohiveFixture as MockAerohive
 
 def test_stop_server(MockAerohive):
     username = "admin"
